requirements.txt
python-telegram-bot[job_queue,webhooks]==20.0
requests
beautifulsoup4
python-dotenv
Flask       
waitress    